# Backend/OnlineEnergyUtilityPlateform/OnlineEnergyUtilityPlateform/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Online Energy Utility Plateform API",
    "version": "v1"
  },
  "paths": {
    "/api/accounts/Delete/{id}": {
      "delete": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserListDtoEntityResponseModel"
            }
          }
        }
      }
    },
...
    "/api/accounts/registration": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserForRegistrationDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringEntityResponseListModel"
            }
          }
        }
      }
    },
    "/api/accounts/login": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserForAuthenticationDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringEntityResponseModel"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "StringEntityResponseListModel": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "StringEntityResponseModel": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "data": { },
        "message": {
          "type": "string"
        }
      }
    },
...
    "UserForAuthenticationDto": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "UserForRegistrationDto": {
      "required": [
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "userName": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        },
        "roleId": {
          "type": "string"
        }
      }
    },
...
    "UserListDtoEntityResponseModel": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "data": { },
        "message": {
          "type": "string"
        }
      }
    }
  }
}